se pulieron detalles en las validaciones asi como se termino el consumo en cursos, roles

+++ src/app/dashboard/roles/roles.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormBuilder, FormControl, Validators } from '@angular/forms';
+import { CookieService } from 'ngx-cookie-service';
 import { Subject } from 'rxjs';
 import Swal from 'sweetalert2';
 import { RolesService } from './services/roles.service';
@@ -20,7 +21,8 @@ export class RolesComponent implements OnInit, OnDestroy {
 
   constructor(
     private _rolesService: RolesService,
-    private _fb: FormBuilder
+    private _fb: FormBuilder,
+    private _cockieService: CookieService,
   ) {
     this.formRol = this._fb.group({
       tipo_rol: new FormControl('', [Validators.required]),
@@ -35,11 +37,27 @@ export class RolesComponent implements OnInit, OnDestroy {
     this.getRoles();
   }
 
+  isEmpty() {
+    if (this.roles.length === 0) {
+      Swal.fire({
+        title: 'Error',
+        text: 'No hay roles registrados, por favor registra un rol',
+        icon: 'error',
+        confirmButtonText: 'Ok'
+      })
+    }
+  }
+
+
   getRoles(){
     this._rolesService.getRoles().subscribe(
       (data : any) => {
         this.roles = data
         console.log(data)
+      },
+      
+      (error) => {
+        this.isEmpty();
       });
   }
 
@@ -56,27 +74,44 @@ export class RolesComponent implements OnInit, OnDestroy {
         console.log(data)
       });
   }
+
+  // if(rol != '1'){
+  //   Swal.fire({
+  //     title: 'Estos roles son por defecto, no se pueden eliminar',
+  //     icon: 'error',
+  //     showConfirmButton: true,
+  //   })
+  // }
+
   eliminarRol(id : any){
-    Swal.fire({
-      title: '¿Estas seguro?',
-      text: "No podras revertir esto",
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#3085d6',
-      cancelButtonColor: '#d33',
-      confirmButtonText: 'Si, Eliminar!'
-    }).then((result) => {
-      if (result.value) {
-        this._rolesService.eliminarRol(id).subscribe();
-        Swal.fire(
-          'Eliminado!',
-          'El rol ha sido eliminado.',
-          'success'
-          ).then(() => {
-            location.reload();
-          })
-      }
-    })
+    if ( id != '1' && id != '2') {
+      Swal.fire({
+        title: '¿Estas seguro?',
+        text: "No podras revertir esto",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, Eliminar!'
+      }).then((result) => {
+        if (result.value) {
+          this._rolesService.eliminarRol(id).subscribe();
+          Swal.fire(
+            'Eliminado!',
+            'El rol ha sido eliminado.',
+            'success'
+            ).then(() => {
+              location.reload();
+            })
+        }
+      })
+    } else {
+      Swal.fire({
+        title: 'Estos roles son por defecto, no se pueden eliminar',
+        icon: 'error',
+        showConfirmButton: true,
+      })
+    }
   }
 
   ngOnDestroy(): void {
